requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
pydantic==2.5.3
Werkzeug==2.3.7
python-dotenv==1.0.0
flask_cors
flask_migrate
psycopg2
pydantic[email]